1.1
implementate requests

+++ resources/views/clients/index.blade.php
@@ -12,39 +12,41 @@
 
                 <div class="card-body">
                     @if (session('status'))
-                        <div class="alert alert-success" role="alert">
+                        <div class="alert alert-{{session('type')}}" role="alert">
                             {{ session('status') }}
                         </div>
                     @endif
 
-                    <table class="table">
-                        <tr>
-                            <th>@lang('clients.ID')</th>
-                            <th>@lang('clients.Name')</th>
-                            <th></th>
-                            <th></th>
-                        </tr>
-                        @foreach ($clients as $key => $value)
+                    @if($clients->count())
+                        <table class="table">
                             <tr>
-                                <td>{{$value->id}}</td>
-                                <td>{{$value->name}}</td>
-                                <td>
-                                    <a href="{{route('clients.edit', [$value->id])}}">@lang('clients.edit')</a>
-                                </td>
-                                <td>
-                                    <a href="javascript:void(0)" onclick="
-                                        $.ajax({
-                                            url: '{{route('clients.destroy', [$value->id]).'?_token='.csrf_token()}}',
-                                            type: 'DELETE',
-                                            success: function(result) {
-                                                location.reload();
-                                            }
-                                        })
-                                    ">@lang('clients.delete')</a>
-                                </td>
+                                <th>@lang('clients.ID')</th>
+                                <th>@lang('clients.Name')</th>
+                                <th></th>
+                                <th></th>
                             </tr>
-                        @endforeach
-                    </table>
+                            @foreach ($clients as $key => $value)
+                                <tr>
+                                    <td>{{$value->id}}</td>
+                                    <td>{{$value->name}}</td>
+                                    <td>
+                                        <a href="{{route('clients.edit', [$value->id])}}">@lang('clients.edit')</a>
+                                    </td>
+                                    <td>
+                                        <a href="javascript:void(0)" onclick="
+                                            $.ajax({
+                                                url: '{{route('clients.destroy', [$value->id]).'?_token='.csrf_token()}}',
+                                                type: 'DELETE',
+                                                success: function(result) {
+                                                    location.reload();
+                                                }
+                                            })
+                                        ">@lang('clients.delete')</a>
+                                    </td>
+                                </tr>
+                            @endforeach
+                        </table>
+                    @endif
                 </div>
             </div>
         </div>
